Change data structure for nodes
Before, it was an array. But searching a single node in that array,
in order to know if it existed, was O(N). Then, I made a Set
property with the id for every node. But then it struck me:
Instead of having two different data structures, I could just store
the nodes in an object. So here it is!